feat: perform `Open` action
Open and display file content. For now, Notepad can only read text
files. So if we open a file not in text format, we will receive `Cannot
open file` warning. Window title will be set to  the file name if the
file was opened succesfully.

Update `New` action to have `*New` window title when clicked.

Issue #2